fix InputDataType for bert-99

+++ program/packed-bert-qaic-loadgen/harness.cpp
@@ -100,7 +100,7 @@ Program::Program() {
         new Benchmark<InCopy, OutCopy, uint64_t, float>(settings, in, out));
   else
     benchmark.reset(
-        new Benchmark<InCopy, OutCopy, uint64_t, uint8_t>(settings, in, out));
+        new Benchmark<InCopy, OutCopy, uint32_t, uint8_t>(settings, in, out));
 
   // create enough ring buffers for each activation within all devices
   ring_buf.resize(settings->qaic_activation_count *
